readme
Members: 1: Georgescu Cristina
Project: Online Shopping Cart for a Music Store

A customer will add one or more product to a shopping card. The purchase will be made after certain business rules, which will make the declarative part of the project. This is done in the framework Drool. Some rules include the case of discounts from the total sum to pay, if one item is not in stock at the moment of the purchase, new customer gets a discount etc.
The interaction in the back ground is done in Java, making the Imperative and Functional part.